streamlit_app/requirements.txt
unidecode==1.3.7
rank-bm25==0.2.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0 
rapidfuzz==3.4.0
gensim==4.3.2
nltk
pandas
pot==0.9.1
streamlit

